requirements.txt
# 파일은 추후 toml 파일로 병합 및 packaging 진행하고 삭제하기!

git+https://github.com/dongseop-kim/universal-datasets.git
torch>=1.13.0
torchvision>=0.14.1
torchmetrics>=1.2.1
albumentations>=1.3.1
hydra-core==1.3.2
timm==0.9.2
wandb==0.16.1
rich==13.7.0
mlflow==2.7.1
pytest
onnx
onnxruntime
